styling of input, buttons in input, placement of branding
