+++ docs/embedding.md
@@ -94,7 +94,6 @@ See it in action:
 | disableHtmlInput                    | boolean                             | false                              |                               |                                       |                                 | If true, strips all html tags out from the input of the user.                                                                                                                                                   |
 | disableInputAutocomplete            | boolean                             | false                              | x                             | x                                     |                                 | If true, disabled the autosuggestions from e.g. mobile touch keyboards.                                                                                                                                         |
 | disableInputAutofocus               | boolean                             | false                              |                               |                                       |                                 | By default, the input will automatically focus when a user opens the widget. If you set this to true, the input will no longer focus when opening the widget.                                                   |
-| disableInputAutogrow                | boolean                             | false                              |                               |                                       |                                 | If true, multiline input is disabled and long messages will scroll horizontally                                                                                                                                 |
 | disableLocalStorage                 | boolean                             | false                              |                               |                                       |                                 | If true, disables storing any information in browsers storage like persistent history and userId. This flag has a higher priority than `useSessionStorage` - setting this to true also disables SessionStorage. |
 | disablePersistentHistory            | boolean                             | false                              |                               |                                       |                                 | If true, disables storing of the chat history into LocalStorage (used for persistence).                                                                                                                         |
 | disableRenderURLsAsLinks            | boolean                             | false                              |                               |                                       |                                 | If true, disables the automatic replacement of URLs in message elements with clickable HTML link elements.                                                                                                                         |
