+++ package.json
@@ -46,7 +46,7 @@
 	"dependencies": {
 		"@braintree/sanitize-url": "^6.0.0",
 		"@cognigy/chat-components": "0.2.8",
-		"@cognigy/socket-client": "5.0.0-beta.2",
+		"@cognigy/socket-client": "5.0.0-beta.4",
 		"@emotion/cache": "^10.0.29",
 		"@emotion/react": "^11.7.1",
 		"@emotion/styled": "^11.6.0",
@@ -67,7 +67,7 @@
 		"react-player": "^2.9.0",
 		"react-redux": "^7.2.6",
 		"react-responsive-carousel": "^3.2.22",
-		"react-textarea-autosize": "^8.3.3",
+		"react-textarea-autosize": "8.5.3",
 		"redux": "^4.0.4",
 		"remarkable": "^2.0.1",
 		"styled-jss": "^2.2.3",
@@ -89,6 +89,7 @@
 		"@babel/preset-react": "^7.16.7",
 		"@babel/preset-typescript": "^7.16.7",
 		"@babel/runtime": "^7.16.7",
+		"@types/dom-speech-recognition": "0.0.4",
 		"@types/dompurify": "^2.0.0",
 		"@types/react": "^16.9.11",
 		"@types/react-dom": "^16.9.4",
